👌 IMPROVE: Code block highlighting (#478)
In markdown a code block (a.k.a fence) is of the form:

````markdown
```language
source text
```
````

MyST-Parser mimics the `code-block` directive to render these blocks:

In sphinx, the lexer name is only recorded as the `language` attribute,
and the text is lexed later by pygments within the `visit_literal_block`
method of the output format ``SphinxTranslator``.
This is the current logic. 

However, in docutils, this directive directly parses the text with the pygments lexer, if syntax highlighting is enabled (the default). This was not handled.

Both cases are now handled, and additionally the following configuration are added:

- `myst_highlight_code_blocks` (docutils only): If True (default) use pygments to create lexical tokens for the given language, otherwise skip lexical analysis
- `myst_number_code_blocks`: A list of languages to add line numbers to