👌 IMPROVE: Allow for heading anchor links in docutils
This aligns the treatment of `[](#target)` style links for docutils with sphinx,
such that they are linked to a heading slug.

The core behaviour for sphinx is not changed,
except that failed reference resolution
now emits a `myst.xref_missing` warning (as opposed to a `std.ref` one), with a clearer warning message.
Also on failure, the reference is still created,
for people who wish to suppress the warning (see e.g. #677)

This is another step towards https://github.com/executablebooks/myst-enhancement-proposals/pull/10