👌 Enhance `[](#id)` references (#707)
This PR moves myst-parser in the direction of https://github.com/executablebooks/myst-enhancement-proposals/pull/10, in a relatively back compatible manner, and in a way that supports both docutils and sphinx.

It expands the capability of `[](#id)` to more than just linking to heading slugs, with the order of specificity being:

1. If it matches a local (to that document) "explicit" `std:ref` target, then link to that and stop
   - Note, currently only `std:ref` domain/type are supported, e.g. not `math` etc. That's more difficult and can come later
2. If it matches a local (to that document) "implicit" heading slug, then link to that and stop
3. If using docutils (i.e. single-page) build, then stop here and emit a `myst.xref_missing` warning
4. If using sphinx then create a `pending_xref` node, and hand-off to sphinx's "any" resolver, which takes effect once all documents have been read:
    - This first tries to resolve against a local (to the project) reference and, if matching, stops
    - Otherwise also try to match against any intersphinx reference
    - Otherwise emit a ~~`myst.ref`~~ `myst.xref_missing`  warning
 
If the text is explicit, e.g. `[text](#id)`, that text is used, otherwise a determination of implicit text is attempted, e.g. based on the section title or figure caption.