Rebuild link labels for Sphinx i18n support
When processing translations in Sphinx, after Sphinx's i8n transform
replaces parts of a document with a translation, a document is processed
a message at a time through the MyST-parser. For references using link
labels, these may not be to the parse for the specific messages that
need them and links may be left unbuilt after processing a document.

To prevent issues when using translations, cache any detected label
references when processing a given document. And when individual
messages are parsed after translation, rebuild any label references and
append them to the message first. This will ensure references will be
resolved/built as expected.

Signed-off-by: James Knight <james.d.knight@live.com>