👌 IMPROVE: External links in new tab (#856)
Add support to enable external links to open in a new tab. Expose the
configuration option myst_links_external_new_tab, which if set will set
all URL links to open in a new tab on the browser.

Besides that, we allow the user to set the target and rel attributes
when using the extension inline_attrs (#820).

Closes #820
Closes #856

Signed-off-by: Marjus Cako <marios.cako@hpe.com>